requirements-dev.txt
pytest
aiosmtpd
PyJWT
